complete mapping configuration
